# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/notifications/intents": {
      "get": {
        "operationId": "listIntents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationIntentListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notifications"]
      }
    },
    "/v1/notifications/intents/{id}": {
      "get": {
        "operationId": "readIntent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationIntent"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notifications"]
      }
    },
    "/v1/notifications/intents/{id}/attempts": {
      "get": {
        "operationId": "listAttempts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationAttemptListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notifications"]
      }
    },
...
  },
  "info": {
    "title": "Carespaces API",
    "description": "Carespaces MVP REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NotificationIntent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "notificationClass",
          "channel",
          "subjectType",
          "subjectId",
          "recipientUserId",
          "recipientRef",
          "bodyRedacted",
          "correlationId",
          "status",
          "attempts",
          "nextAttemptAt",
          "deliveredAt",
          "terminalFailedAt",
          "lastError",
          "acknowledgedAt",
          "opsTaskId",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "notificationClass": {
            "type": "string",
            "enum": [
              "incident_ack",
              "sos",
              "credential_expiry_block",
              "replacement_failed",
              "shift_reminder",
              "reservation_expiry",
              "payment_expiry",
              "customer_approval_reminder",
              "dispute_update",
              "payout_retry",
              "system"
            ]
          },
          "channel": {
            "type": "string",
            "enum": ["push", "sms", "email", "in_app"]
          },
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "recipientUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "recipientRef": {
            "type": "string"
          },
          "bodyRedacted": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "LEASED",
              "DELIVERED",
              "TERMINAL_FAILED",
              "CANCELLED"
            ]
          },
          "attempts": {
            "type": "integer",
            "minimum": 0
          },
          "nextAttemptAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "terminalFailedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastError": {
            "type": "string",
            "nullable": true
          },
          "acknowledgedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opsTaskId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationAttempt": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "intentId",
          "attemptNumber",
          "channel",
          "adapterName",
          "status",
          "providerMessageRef",
          "errorClass",
          "errorMessage",
          "startedAt",
          "completedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "intentId": {
            "type": "string",
            "format": "uuid"
          },
          "attemptNumber": {
            "type": "integer",
            "minimum": 1
          },
          "channel": {
            "type": "string",
            "enum": ["push", "sms", "email", "in_app"]
          },
          "adapterName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["FIRED", "FAILED", "DEAD_LETTER"]
          },
          "providerMessageRef": {
            "type": "string",
            "nullable": true
          },
          "errorClass": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "NotificationIntentListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["intents", "generatedAt"],
        "properties": {
          "intents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationIntent"
            }
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationAttemptListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["intentId", "attempts", "generatedAt"],
        "properties": {
          "intentId": {
            "type": "string",
            "format": "uuid"
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationAttempt"
            }
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}